refactor: extract shared/github.ts for gh CLI operations (#31)
Move fetchOpenIssues, createSubIssues, createPR from orchestrator and
discoverMergeablePRs, mergePR from finalize into shared/github.ts.
Co-locate GitHubIssue, MergeOrder, MergeStrategy types. Both tools
import from shared; re-exports maintain backward compatibility.
