+++ tools/finalize/types.ts
@@ -2,7 +2,10 @@
  * Types for the finalize (PR merge) tool.
  */
 
-export type MergeStrategy = 'squash' | 'merge' | 'rebase';
+import type { MergeStrategy } from '../../shared/github.ts';
+
+// Re-exported from shared for backward compatibility
+export type { MergeStrategy, MergeOrder } from '../../shared/github.ts';
 
 export type PRStatus = 'pending' | 'merged' | 'failed' | 'conflict' | 'skipped';
 
@@ -34,10 +37,3 @@ export interface FinalizeState {
 }
 
 export type { ConflictInfo } from '../../shared/git.ts';
-
-export interface MergeOrder {
-	issueNumber: number;
-	prNumber: number;
-	branch: string;
-	baseBranch: string;
-}
